+++ src/main/java/plana/replan/domain/goal/dto/refine/GoalRefinementResponseDto.java
@@ -1,4 +1,4 @@
-package plana.replan.domain.goal.dto;
+package plana.replan.domain.goal.dto.refine;
 
 import io.swagger.v3.oas.annotations.media.Schema;
