+++ src/main/java/plana/replan/domain/goal/service/GoalAiService.java
@@ -14,15 +14,15 @@ import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClient;
 import org.springframework.web.client.RestClientException;
-import plana.replan.domain.goal.dto.GoalRefinementRequestDto;
-import plana.replan.domain.goal.dto.GoalRefinementResponseDto;
-import plana.replan.domain.goal.dto.NoteItemDto;
-import plana.replan.domain.goal.dto.RecommendedTodoDto;
-import plana.replan.domain.goal.dto.RefinedDeadline;
-import plana.replan.domain.goal.dto.RefinedField;
-import plana.replan.domain.goal.dto.RefinedNotes;
-import plana.replan.domain.goal.dto.TodoRecommendationRequestDto;
-import plana.replan.domain.goal.dto.TodoRecommendationResponseDto;
+import plana.replan.domain.goal.dto.recommend.RecommendedTodoDto;
+import plana.replan.domain.goal.dto.recommend.TodoRecommendationRequestDto;
+import plana.replan.domain.goal.dto.recommend.TodoRecommendationResponseDto;
+import plana.replan.domain.goal.dto.refine.GoalRefinementRequestDto;
+import plana.replan.domain.goal.dto.refine.GoalRefinementResponseDto;
+import plana.replan.domain.goal.dto.refine.NoteItemDto;
+import plana.replan.domain.goal.dto.refine.RefinedDeadline;
+import plana.replan.domain.goal.dto.refine.RefinedField;
+import plana.replan.domain.goal.dto.refine.RefinedNotes;
 import plana.replan.domain.goal.exception.GoalErrorCode;
 import plana.replan.global.exception.CustomException;
